fix: patch schemas at runtime
